Create education
In this code we use the "sklearn" library to train the model. We download data on prices per square meter of real estate from a real estate website. Next, we display a price map to give you an approximate range of cents per square meters. After all this, we load the data into the training module. As a result, we get a model where we indicate square meters and get their price. Finally, for convenience, we save the new column into an excel file.